+++ src/App/components/FileSelector.js
@@ -6,8 +6,8 @@ const { dialog } = remote;
 
 const FileSelector = ({
   save = false,
+  disabled = false,
   file,
-  label,
   options,
   onFileSelected,
 }) => {
@@ -21,9 +21,8 @@ const FileSelector = ({
   };
   return (
     <div className='file-selector'>
-      <div className='file-selector__label'>{label}</div>
       <div className='file-selector__button'>
-        <Button text='Select' onClick={selectFile} />
+        <Button text='Select' onClick={selectFile} disabled={disabled} />
       </div>
       <div className='file-selector__filename'>{file}</div>
     </div>