+++ src/App/components/cards/index.js
@@ -2,21 +2,19 @@ import React from 'react';
 
 import TextAndAudioCard from './TextAndAudioCard';
 // import TimingCard from './TimingCard';
-import BackgroundCard from './BackgroundCard';
+import DisplayCard from './DisplayCard';
 import OutputCard from './OutputCard';
-import FontCard from './FontCard';
 
 export {
   TextAndAudioCard,
   // TimingCard,
-  BackgroundCard,
+  DisplayCard,
   OutputCard,
-  FontCard,
 };
 
 export const cards = [
   {
-    title: 'Text and Audio',
+    title: 'HearThis',
     description:
       'This first step is where you select the HearThis project, book and chapter.' +
       ' This folder should contain the text and audio files that will be used in the video.',
@@ -29,20 +27,14 @@ export const cards = [
   //   content: <TimingCard />,
   // },
   {
-    title: 'Background',
+    title: 'Display',
     description:
-      'Then select a background image or video that will be used as the background of the generated video',
-    content: <BackgroundCard />,
-  },
-  {
-    title: 'Font',
-    description:
-      'Then select the font to use for the text in the generated video',
-    content: <FontCard />,
+      'Now configure what the video will look like.',
+    content: <DisplayCard />,
   },
   {
     title: 'Output',
-    description: "Finally, select where you'll save the generated video",
+    description: "Finally, select where you'll save the generated video.",
     content: <OutputCard />,
   },
 ];