+++ src/App/index.js
@@ -52,17 +52,18 @@ class App extends React.PureComponent {
   openOutputFolder = () => {
     const { outputFile } = this.props.store;
     ipcRenderer.send('open-output-folder', outputFile);
-    this.setState({ status: AppStatus.configuring });
+    this.reset();
   };
 
   onStart = () => {
     this.setState({ status: AppStatus.processing }, () => {
-      const { hearThisFolder, backgroundFile, font, outputFile } = this.props.store;
+      const { hearThisFolder, background, text, speechBubble, outputFile } = this.props.store;
       const args = {
         hearThisFolder,
-        backgroundFile,
-        font,
-        outputFile,
+        background,
+        text,
+        speechBubble,
+        output: outputFile,
       };
       console.log('Requesting processing', args);
       ipcRenderer.send('did-start-conversion', args);
@@ -87,7 +88,7 @@ class App extends React.PureComponent {
     switch(status) {
       case AppStatus.done:
         return (
-          <div>
+          <div className="app__footer-success">
             <p className={Classes.TEXT_LARGE}>Your Bible Karaoke video has been created!</p>
             <ActionButton
               large
@@ -95,6 +96,7 @@ class App extends React.PureComponent {
               text='Open output folder'
               onClick={this.openOutputFolder}
             />
+            <ActionButton large onClick={this.reset} text='Make another video...' />
           </div>
         );
       case AppStatus.error: