+++ CHANGELOG.md
@@ -6,6 +6,19 @@
 * :thumbsup: Improvement
 * :boom: Bug fix
 
+## v0.3.0
+
+(Under the hood, the CLI code is now integrated into this repo, rather than being a npm dependency)
+
+* :star: Lots of new display configuration options:
+    * Use a solid background color as an alternative to an image
+    * Set the font size and color as well as font family
+    * Set the highlight color
+    * Set the 'speech bubble' color and opacity
+* :thumbsup: Extra 'Make another video...' button to reset the UI after a video finishes.
+* :thumbsup: Added a refresh button next to the HearThis project selection
+* :boom: Fixed background color - fills window when window is resized
+
 ## v0.2.2
 
 * :boom: Corrected chapter numbering. Chapter '0' is now labeled as 'Intro'.